refactor out string utils into module file
